refactor(backend): group app bootstrap and infrastructure files under core/
Move app.ts, container.ts, container-tokens.ts, ioc.ts, and middlewares/
into a new core/ directory to separate infrastructure wiring from domain
modules. Update all import paths and tsoa.json accordingly.
