+++ backend/CLAUDE.md
@@ -14,16 +14,16 @@ Flat structure — no `src/` directory. Nest-style domain modules under `modules
   - `*.module.ts` — binds this domain's classes into the Inversify container
 - `types/` — shared types exported from `types/index.ts`
 - `prisma/` — schema and migrations
-- `container.ts` / `container-tokens.ts` / `ioc.ts` — DI wiring (see below)
+- `core/` — app bootstrap and infrastructure: `app.ts`, `container.ts`, `container-tokens.ts`, `ioc.ts`, `middlewares/`
 
 Domains: `auth/`, `files/`, `projects/`, `claude/`. The `claude` module holds the WebSocket controller (`claude-ws.controller.ts`) and Claude Code process management; unlike REST modules, its controller is resolved manually in `backend/index.ts` since tsoa does not handle WebSocket routes.
 
 ## Dependency Injection
 
-Inversify. Classes declare dependencies with `@injectable()` and `@inject(...)`; each domain's `*.module.ts` registers its bindings into the shared `container` from `container.ts`.
+Inversify. Classes declare dependencies with `@injectable()` and `@inject(...)`; each domain's `*.module.ts` registers its bindings into the shared `container` from `core/container.ts`.
 
-- `container-tokens.ts` holds shared symbol tokens (e.g. `PRISMA_CLIENT`) so modules can import tokens without pulling in the full container graph (avoids a cycle: container → module → repository → container).
-- `ioc.ts` exposes the container to tsoa via `iocContainer.get`. `tsoa.json` points to it with `"iocModule": "./ioc"`, so tsoa resolves controllers through Inversify.
+- `core/container-tokens.ts` holds shared symbol tokens (e.g. `PRISMA_CLIENT`) so modules can import tokens without pulling in the full container graph (avoids a cycle: container → module → repository → container).
+- `core/ioc.ts` exposes the container to tsoa via `iocContainer.get`. `tsoa.json` points to it with `"iocModule": "./core/ioc"`, so tsoa resolves controllers through Inversify.
 - To add a new dependency: decorate the class with `@injectable()`, bind it in the appropriate `*.module.ts`, and inject it via constructor parameter.
 
 ## tsoa
@@ -40,7 +40,7 @@ SQLite via Prisma + libsql adapter. Schema: `prisma/schema.prisma`.
 
 - Default DB path: `~/.lgtmai/lgtmai.db` (override with `DB_PATH` env var)
 - Dev: `prisma migrate dev` | Prod: `prisma migrate deploy`
-- `PrismaClient` is bound in `container.ts` under the `PRISMA_CLIENT` token; repositories inject it with `@inject(PRISMA_CLIENT)`.
+- `PrismaClient` is bound in `core/container.ts` under the `PRISMA_CLIENT` token; repositories inject it with `@inject(PRISMA_CLIENT)`.
 
 ## Conventions
