+++ backend/core/middlewares/errorHandler.ts
@@ -1,6 +1,6 @@
 import { Request, Response, NextFunction } from 'express';
 import { ValidateError } from '@tsoa/runtime';
-import { AppError } from '../errors/AppError.js';
+import { AppError } from '../../errors/AppError.js';
 
 export function errorHandler(
   err: unknown,
